requirements.txt
altgraph==0.17
asgiref==3.3.4
beautifulsoup4~=4.9.3
bs4==0.0.1
certifi==2020.12.5
cssselect~=1.1.0
coverage==5.5
Django==3.2.3
django-widget-tweaks==1.4.8
django-extensions==3.1.3
flake8==3.9.2
future==0.18.2
importlib-metadata==4.0.1
joblib==1.0.1
keras-tuner==1.0.3
matplotlib~=3.3.2
mccabe==0.6.1
mysqlclient==2.0.3
numpy~=1.19.5
opencv-python==4.5.2.52
pandas==1.2.4
pefile==2021.5.13
pycodestyle==2.7.0
pyflakes==2.3.1
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pylint==2.4.4
pyproj==3.0.1
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six~=1.15.0
selenium~=3.141.0
six~=1.15.0
sqlparse==0.4.1
tensorflow==2.5.0
threadpoolctl==2.1.0
typing-extensions~=3.7.4
zipp==3.4.1
